cloud-init: Add UserData:MarshalMultiPart helper
This will allow better separation between the high level decisions
(usermode networking enabled/disabled and --cloud-init used or not) and
the low-level byte details
