+++ pkg/machine/cloudinit/cloudinit_windows.go
@@ -3,11 +3,9 @@
 package cloudinit
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
-	"mime/multipart"
 	"net/http"
 
 	"github.com/containers/podman/v5/pkg/machine/hyperv/hutil"
@@ -69,15 +67,9 @@ func generateUserData(mc *vmconfigs.MachineConfig) ([]byte, error) {
 		}
 	}
 
-	internalUserDataBytes, err := internalUserData.Marshal()
-	if err != nil {
-		logrus.Errorf("Error marshaling to YAML: %v", err)
-		return nil, err
-	}
-
 	// If user has not provided any custom user-data, return the generated one
 	if mc.CloudInitConfig.UserData == nil {
-		return internalUserDataBytes, nil
+		return internalUserData.Marshal()
 	}
 
 	// if user has provided a custom user-data but we're not on Hyper-V/user-mode networking, return it as-is
@@ -93,30 +85,7 @@ func generateUserData(mc *vmconfigs.MachineConfig) ([]byte, error) {
 	// we need to merge our generated user data with user's one
 	// To do it we create a MIME multi-part archive
 	// with both files
-	buf := new(bytes.Buffer)
-	writer := multipart.NewWriter(buf)
-
-	// add our configuration as the first part
-	if err := createCloudConfigPart(writer, internalUserDataBytes); err != nil {
-		return nil, fmt.Errorf("failed to create internal cloud-config part: %w", err)
-	}
-
-	// Add the user's config as a second part
-	if err := createCloudConfigPart(writer, userUserData); err != nil {
-		return nil, fmt.Errorf("failed to create user cloud-config part: %w", err)
-	}
-
-	if err := writer.Close(); err != nil {
-		return nil, fmt.Errorf("failed to close multipart writer: %w", err)
-	}
-
-	// finalize mime archive with top-level header
-	finalContent := new(bytes.Buffer)
-	topLevelHeader := fmt.Sprintf("Content-Type: multipart/mixed; boundary=%s\nMIME-Version: 1.0\n\n", writer.Boundary())
-	finalContent.WriteString(topLevelHeader)
-	finalContent.Write(buf.Bytes())
-
-	return finalContent.Bytes(), nil
+	return internalUserData.MarshalMultiPart(userUserData)
 }
 
 func getGvForwarderBytes() ([]byte, error) {
