+++ pkg/machine/cloudinit/userdata.go
@@ -1,6 +1,11 @@
 package cloudinit
 
 import (
+	"bytes"
+	"fmt"
+	"mime/multipart"
+	"net/textproto"
+
 	"github.com/containers/podman/v5/pkg/machine"
 	"github.com/containers/podman/v5/pkg/machine/vmconfigs"
 	"gopkg.in/yaml.v3"
@@ -63,4 +68,52 @@ func (userData *UserData) Marshal() ([]byte, error) {
 	data = append([]byte(headerLine), data...)
 
 	return data, nil
+}
+
+func (userData *UserData) MarshalMultiPart(extraData []byte) ([]byte, error) {
+	buf := new(bytes.Buffer)
+	writer := multipart.NewWriter(buf)
+
+	userDataBytes, err := userData.Marshal()
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshall user-data: %w", err)
+	}
+
+	// add our configuration as the first part
+	if err := createCloudConfigPart(writer, userDataBytes); err != nil {
+		return nil, fmt.Errorf("failed to create internal cloud-config part: %w", err)
+	}
+
+	// Add the user's config as a second part
+	if err := createCloudConfigPart(writer, extraData); err != nil {
+		return nil, fmt.Errorf("failed to create user cloud-config part: %w", err)
+	}
+
+	if err := writer.Close(); err != nil {
+		return nil, fmt.Errorf("failed to close multipart writer: %w", err)
+	}
+
+	// finalize mime archive with top-level header
+	finalContent := new(bytes.Buffer)
+	topLevelHeader := fmt.Sprintf("Content-Type: multipart/mixed; boundary=%s\nMIME-Version: 1.0\n\n", writer.Boundary())
+	finalContent.WriteString(topLevelHeader)
+	finalContent.Write(buf.Bytes())
+
+	return finalContent.Bytes(), nil
+}
+
+func createCloudConfigPart(writer *multipart.Writer, content []byte) error {
+	header := textproto.MIMEHeader{}
+	// Set the specific Content-Type that cloud-init recognizes for configuration files
+	header.Set("Content-Type", "text/cloud-config")
+
+	partWriter, err := writer.CreatePart(header)
+	if err != nil {
+		return fmt.Errorf("failed to create new MIME part: %w", err)
+	}
+
+	if _, err := partWriter.Write(content); err != nil {
+		return fmt.Errorf("failed to write content to MIME part: %w", err)
+	}
+	return nil
 }
